requirements.txt
streamlit
requests
fpdf2
transformers
torch
